vieles

+++ Teil_I/1_Lets-go/1_10-signed-unsigned-char.cpp
@@ -12,8 +12,8 @@ int main(){
     }
 
     std::cout << "Grenzwerte für char in int umgewandelt: \n";
-    std::cout << "Minimum = " << std::numeric_limits<char>::min() <<'\n';
-    std::cout << "Maximum = " << std::numeric_limits<char>::max() <<'\n';
+    std::cout << "Minimum = " << static_cast<int>(std::numeric_limits<char>::min()) <<'\n';
+    std::cout << "Maximum = " << static_cast<int>(std::numeric_limits<char>::max()) <<'\n';
     std::cout << "Maximum = " << std::numeric_limits<int>::max() <<'\n';
 
     return 0;
